slash/invite-all: refactoring
